requirements.txt
selenium 
python-telegram-bot
python-telegram-bot[job-queue]
pymongo[srv]
matplotlib